# docs/swagger.yaml
basePath: /v1
definitions:
...
  pkgs.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
  role.AssignPermissionsReq:
    properties:
      permission_ids:
        items:
          type: string
        minItems: 1
        type: array
    required:
    - permission_ids
    type: object
  role.CreateRoleReq:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - name
    type: object
  role.RoleListRes:
    properties:
      list:
        items:
          $ref: '#/definitions/role.RoleRes'
        type: array
      total:
        type: integer
    type: object
  role.RoleRes:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  role.UpdateRoleReq:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: https://www.swagger.io/support
  description: This is a sample server for go-pg-demo server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://swagger.io/terms/
  title: Go-PG Demo API
  version: "1.0"
paths:
...
  /role:
    post:
      consumes:
      - application/json
      description: 创建一个新角色
      parameters:
      - description: 创建角色的请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/role.CreateRoleReq'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功，返回角色ID
          schema:
            allOf:
            - $ref: '#/definitions/pkgs.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/pkgs.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/pkgs.Response'
      summary: 创建角色
      tags:
      - role
  /role/{id}:
    delete:
      consumes:
      - application/json
      description: 根据ID删除一个角色
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功，返回影响行数
          schema:
            allOf:
            - $ref: '#/definitions/pkgs.Response'
            - properties:
                data:
                  format: int64
                  type: integer
              type: object
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/pkgs.Response'
      summary: 删除角色
      tags:
      - role
    get:
      consumes:
      - application/json
      description: 根据ID获取单个角色的详细信息
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功，返回角色信息
          schema:
            allOf:
            - $ref: '#/definitions/pkgs.Response'
            - properties:
                data:
                  $ref: '#/definitions/role.RoleRes'
              type: object
        "404":
          description: 角色不存在
          schema:
            $ref: '#/definitions/pkgs.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/pkgs.Response'
      summary: 获取角色详情
      tags:
      - role
    put:
      consumes:
      - application/json
      description: 根据ID更新角色的名称或描述
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新角色的请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/role.UpdateRoleReq'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/pkgs.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/pkgs.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/pkgs.Response'
      summary: 更新角色信息
      tags:
      - role
  /role/{id}/permission:
    post:
      consumes:
      - application/json
      description: 清空角色现有权限，并重新关联新的权限列表
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: string
      - description: 分配权限的请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/role.AssignPermissionsReq'
      produces:
      - application/json
      responses:
        "200":
          description: 分配成功
          schema:
            $ref: '#/definitions/pkgs.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/pkgs.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/pkgs.Response'
      summary: 为角色分配权限
      tags:
      - role
  /role/list:
    get:
      consumes:
      - application/json
      description: 根据条件分页获取角色列表
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      - description: 按名称模糊查询
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功，返回角色列表和总数
          schema:
            allOf:
            - $ref: '#/definitions/pkgs.Response'
            - properties:
                data:
                  $ref: '#/definitions/role.RoleListRes'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/pkgs.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/pkgs.Response'
      summary: 分页查询角色
      tags:
      - role
...
securityDefinitions:
  JWT:
    description: JWT token for authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"